requirements.txt
pip
sqlalchemy
Flask
prometheus_client
prometheus_flask_exporter
pyserial